fix: address PR review feedback for quota predicates and CLI flags
Move IsNotTerminating/IsLiveForQuota to pkg/utils and rename quota Redis
breaker flags for clarity per review.
